requirements.txt
# Django & DRF
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-tenants==3.10.0
django-cors-headers==4.6.0
django-filter==24.3

# Database
psycopg2-binary==2.9.10

# Environment & Configuration
python-dotenv==1.0.1

# Background Tasks
celery==5.4.0
redis==5.2.1

# Image Handling
Pillow>=11.1.0
